+++ GameEngines/UltiTTT/__init__.py
@@ -1,3 +1,14 @@
+import GameEngines
+from GameEngines.UltiTTT import *
+
 from .utilsTypes import Move, Coords
 
-from .BoardState import BoardState
+try:
+    __doc__ = GameEngines.UltiTTT.__doc__
+    if hasattr(GameEngines.UltiTTT, "__all__"):
+        __all__ = GameEngines.UltiTTT.__all__
+
+    from .BoardState import BoardState
+
+except AttributeError:
+    from .PythonEngine.BoardState import BoardState
