+++ GameEngines/UltiTTT/__init__.pyi
@@ -1,0 +1,4 @@
+from .utilsTypes import *
+from .BoardState import BoardState as BoardState
+
+from .UltiTTT import *
